[Backport] CVE-2026-19167: Integer overflow in GPU
Manual cherry-pick. Originally reviewed on
https://chromium-review.googlesource.com/c/chromium/src/+/8127906

[M151] [GPU Security] Widen CompoundImageBacking content id to 64-bit.

Original change's description:
> [GPU Security] Widen CompoundImageBacking content id to 64-bit.
>
> CompoundImageBacking uses content_id_ (initialized to 1) to track which
> element holds the latest content, with 0 serving as the sentinel value
> for uninitialized elements. If content_id_ wraps around 32-bit uint
> space, it collides with 0, causing untouched GPU sub-backings to be
> incorrectly reported as having current content.
>
> Widen latest_content_id_ and ElementHolder::content_id_ from uint32_t to
> uint64_t so overflow is impossible in practice. Add a regression test
> verifying that crossing the 32-bit boundary does not skip the required
> shm->GPU upload on first access.
>
> Bug: 536666274
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8127906
> Commit-Queue: vikas soni <vikassoni@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1666452}

(cherry picked from commit 75e83658911b277adcea35242890db6ffae7f371)

Bug: 538009470,536666274
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8163382
Reviewed-on: https://codereview.qt-project.org/c/qt/qtwebengine-chromium/+/759994

+++ chromium/gpu/command_buffer/service/shared_image/compound_image_backing.h
@@ -166,7 +166,7 @@ class GPU_GLES2_EXPORT CompoundImageBacking : public SharedImageBacking {
     SharedImageBacking* GetBacking();
 
     AccessStreamSet access_streams;
-    uint32_t content_id_ = 0;
+    uint64_t content_id_ = 0;
 
     CreateBackingCallback create_callback;
     std::unique_ptr<SharedImageBacking> backing;
@@ -285,7 +285,8 @@ class GPU_GLES2_EXPORT CompoundImageBacking : public SharedImageBacking {
   // factory from any thread.
   scoped_refptr<SharedImageFactoryRef> shared_image_factory_;
 
-  uint32_t latest_content_id_ = 1;
+  // 64-bit so it never wraps back to a stale element's content id in practice.
+  uint64_t latest_content_id_ = 1;
 
   // Holds all of the "element" backings that make up this compound backing. For
   // each there is a backing, set of streams and tracking for latest content.
